deployment/google_cloud_run/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
google-cloud-storage
python-dotenv
mlflow
